week2/10_code/requirements.txt
langchain==1.2.0
langchain-openai==1.1.6
python-dotenv==1.2.1
gradio==6.2.0
langchain-community==0.4.1